requirements-webapp.txt
Django==1.6.5
django-model-utils
django-haystack
Whoosh
